Implementação da pagina de login e loginBean

+++ src/main/java/br/com/newschool/core/base/ObjectBase.java
@@ -14,14 +14,14 @@ public class ObjectBase implements Serializable {
 	private static final long serialVersionUID = -6064377065615869946L;
 	
 	// Attributes
-	private long id;
+	private int id;
 	
 	// Builds 
 	public ObjectBase() {
 		
 	}
 	
-	public ObjectBase(long id) {
+	public ObjectBase(int id) {
 		this.id = id;
 	}
 
@@ -30,7 +30,7 @@ public class ObjectBase implements Serializable {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
